+++ plugins/importExport/src/FlatImportV2.php
@@ -866,7 +866,7 @@ class FlatImportV2 extends FlatBackup
             case 'post':
                 $line->substitute('post_titre', 'post_title');
                 $line->post_title         = Html::decodeEntities($line->post_title);
-                $line->post_url           = date('Y/m/d/', strtotime($line->post_dt)) . $line->post_id . '-' . $line->post_titre_url;
+                $line->post_url           = date('Y/m/d/', (int) strtotime($line->post_dt)) . $line->post_id . '-' . $line->post_titre_url;
                 $line->post_url           = substr($line->post_url, 0, 255);
                 $line->post_format        = $line->post_content_wiki == '' ? 'xhtml' : 'wiki';
                 $line->post_content_xhtml = $line->post_content;
