+++ src/Core/Backend/Listing/ListingMedia.php
@@ -216,7 +216,7 @@ class ListingMedia extends Listing
             if (!$file->d) {
                 $lst .= '<li>' . ($file->media_priv ? '<img class="media-private" src="images/locker.png" alt="' . __('private media') . '">' : '') . $file->media_title . '</li>' .
                 '<li>' .
-                '<time datetime="' . Date::iso8601(strtotime($file->media_dtstr), App::auth()->getInfo('user_tz')) . '">' .
+                '<time datetime="' . Date::iso8601((int) strtotime($file->media_dtstr), App::auth()->getInfo('user_tz')) . '">' .
                 $file->media_dtstr .
                 '</time>' .
                 ' - ' .
@@ -241,7 +241,7 @@ class ListingMedia extends Listing
                 '<br />' . ($file->d ? '' : ($file->media_priv ? '<img class="media-private" src="images/locker.png" alt="' . __('private media') . '">' : '') . $file->media_title) . '</td>';
             $res .= '<td class="nowrap count">' . (
                 $file->d ? '' :
-                '<time datetime="' . Date::iso8601(strtotime($file->media_dtstr), App::auth()->getInfo('user_tz')) . '">' .
+                '<time datetime="' . Date::iso8601((int) strtotime($file->media_dtstr), App::auth()->getInfo('user_tz')) . '">' .
                 $file->media_dtstr .
                 '</time>'
             ) . '</td>';
